+++ Framework/Source/Revisioning/CDERevisionManager.m
@@ -183,31 +183,39 @@
 - (BOOL)checkIntegrationPrequisitesForEvents:(NSArray *)events error:(NSError * __autoreleasing *)error
 {
     __block BOOL result = YES;
+    __block NSError *outsideBlockError = nil;
     [eventManagedObjectContext performBlockAndWait:^{
+        NSError *localError = nil;
         if (![self checkAllDataFilesExistForStoreModificationEvents:events]) {
-            if (error) *error = [NSError errorWithDomain:CDEErrorDomain code:CDEErrorCodeMissingDataFiles userInfo:nil];
+            localError = [NSError errorWithDomain:CDEErrorDomain code:CDEErrorCodeMissingDataFiles userInfo:nil];
+            outsideBlockError = localError;
             result = NO;
             return;
         }
         
         if (![self checkAllDependenciesExistForStoreModificationEvents:events]) {
-            if (error) *error = [NSError errorWithDomain:CDEErrorDomain code:CDEErrorCodeMissingDependencies userInfo:nil];
+            localError = [NSError errorWithDomain:CDEErrorDomain code:CDEErrorCodeMissingDependencies userInfo:nil];
+            outsideBlockError = localError;
             result = NO;
             return;
         }
         
         if (![self checkContinuityOfStoreModificationEvents:events]) {
-            if (error) *error = [NSError errorWithDomain:CDEErrorDomain code:CDEErrorCodeDiscontinuousRevisions userInfo:nil];
+            localError = [NSError errorWithDomain:CDEErrorDomain code:CDEErrorCodeDiscontinuousRevisions userInfo:nil];
+            outsideBlockError = localError;
             result = NO;
             return;
         }
         
         if (![self checkModelVersionsOfStoreModificationEvents:events]) {
-            if (error) *error = [NSError errorWithDomain:CDEErrorDomain code:CDEErrorCodeUnknownModelVersion userInfo:nil];
+            localError = [NSError errorWithDomain:CDEErrorDomain code:CDEErrorCodeUnknownModelVersion userInfo:nil];
+            outsideBlockError = localError;
             result = NO;
             return;
         }
     }];
+    
+    if (error) *error = outsideBlockError;
     
     return result;
 }