+++ Framework/Tests/CDEPersistentStoreEnsembleMergeTests.m
@@ -1,3 +1,4 @@
+
 //
 //  CDEPersistentStoreEnsembleMergeTests.m
 //  Ensembles
@@ -99,6 +100,8 @@
 
 - (void)tearDown
 {
+    [ensemble1 dismantle];
+    [ensemble2 dismantle];
     didSaveInfo = nil;
     [[NSFileManager defaultManager] removeItemAtPath:rootTestDir error:NULL];
     [super tearDown];